+++ FlintCore/Constraints/Platform.swift
@@ -1,0 +1,16 @@
+//
+//  Platform.swift
+//  FlintCore
+//
+//  Created by Marc Palmer on 01/05/2018.
+//  Copyright © 2018 Montana Floss Co. Ltd. All rights reserved.
+//
+
+import Foundation
+
+public enum Platform: Hashable, Equatable {
+    case iOS
+    case watchOS
+    case tvOS
+    case macOS
+}
