+++ FlintCoreTests/Mocks and Helpers/MockUserToggles.swift
@@ -7,16 +7,30 @@
 //
 
 import Foundation
-import FlintCore
+@testable import FlintCore
 
 class MockUserToggles: UserFeatureToggles {
     var toggles: [FeaturePath:Bool] = [:]
     
-    func isEnabled(_ feature: ConditionalFeatureDefinition.Type) -> Bool {
-        return toggles[feature.identifier] ?? false
+    private var observers = ObserverSet<UserFeatureTogglesObserver>()
+
+    func addObserver(_ observer: UserFeatureTogglesObserver) {
+        let queue = SmartDispatchQueue(queue: .main, owner: self)
+        observers.add(observer, using: queue)
+    }
+    
+    func removeObserver(_ observer: UserFeatureTogglesObserver) {
+        observers.remove(observer)
+    }
+
+    func isEnabled(_ feature: ConditionalFeatureDefinition.Type) -> Bool? {
+        return toggles[feature.identifier]
     }
     
     func setEnabled(_ feature: ConditionalFeatureDefinition.Type, enabled: Bool) {
         toggles[feature.identifier] = enabled
+        observers.notifySync { observer in
+            observer.userFeatureTogglesDidChange()
+        }
    }
 }
